fastapi_app/requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
aiokafka
prometheus-fastapi-instrumentator
python-json-logger